% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tab.R
\name{acond}
\alias{acond}
\alias{aprop}
\title{Transform a frequency table into relative frequencies relative to a margin.}
\usage{
acond(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")

aprop(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")
}
\arguments{
\item{x}{}

\item{MARGIN}{}

\item{all.label}{}

\item{total.margin}{}
}
\description{
Transform a frequency table into relative frequencies relative to a margin.
}

