% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{change}
\alias{change}
\title{Replace elements of x with correspondingly named elements of ll}
\usage{
change(x, ll)
}
\arguments{
\item{x}{}

\item{ll}{}
}
\description{
Replace elements of x with correspondingly named elements of ll
}

