% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConjComp.R
\name{ConjComp}
\alias{ConjComp}
\title{Conjugate complement of span(X) in span(Z) with respect to inner product ip}
\usage{
ConjComp(X, Z = diag(nrow(X)), ip = diag(nrow(X)), tol = 1e-07)
}
\arguments{
\item{X}{}

\item{Z}{}

\item{ip}{}

\item{tol}{}
}
\description{
Conjugate complement of span(X) in span(Z) with respect to inner product ip
}

