% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/up.R
\name{constant}
\alias{constant}
\alias{constant.data.frame}
\alias{constant.default}
\title{Generic test for constant}
\usage{
constant(x, ...)

\method{constant}{default}(x, na.rm = FALSE, ...)

\method{constant}{data.frame}(x, id, all = FALSE, ...)
}
\arguments{
\item{x}{}

\item{na.rm}{(default FALSE)}

\item{id}{}

\item{all}{}

\item{\dots}{}

\item{\dots}{}
}
\description{
Generic test for constant
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{data.frame}: 
}}

