% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Lall}
\alias{Lall}
\title{Hypothesis matrix to test for lmer objects}
\usage{
Lall(fit, nam)
}
\arguments{
\item{fit}{}

\item{pat}{}
}
\value{
hypothesis matrix
}
\description{
Hypothesis matrix to test for lmer objects
}

