% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Lc}
\alias{Lc}
\title{Hypothesis matrix for lmer objects: comparisons with reference level}
\usage{
Lc(fit, nam, ref = 1, verbose = 0)
}
\arguments{
\item{fit}{}

\item{nam}{}

\item{ref}{}

\item{verbose}{}
}
\description{
Hypothesis matrix for lmer objects: comparisons with reference level
}

