% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Lcall}
\alias{Lcall}
\title{Construct hypothesis matrix to test ????}
\usage{
Lcall(fit, factors = getFactorNames(fit), debug = F)
}
\arguments{
\item{fit}{}

\item{factors}{}
}
\value{
a hypothesis matrix
}
\description{
Construct hypothesis matrix to test ????
}

