% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff}
\alias{Ldiff}
\title{Hypothesis matrix to test differencs in factor levels}
\usage{
Ldiff(fit, pat, levnames = c(reflevel, substring(rownames(L), cut + 1)),
  reflevel = "<ref>", cut = nchar(pat), verbose = F)
}
\arguments{
\item{fit}{}

\item{pat}{}

\item{levnames}{}

\item{reflevel}{}

\item{cut}{}

\item{verbose}{}
}
\value{
hypothesis matrix
}
\description{
Hypothesis matrix to test differencs in factor levels
}

