% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff.rdc}
\alias{Ldiff.rdc}
\title{Version of Ldiff used in RDC}
\usage{
Ldiff.rdc(fit, nam, ref = "no longer used")
}
\arguments{
\item{fit}{}

\item{nam}{}

\item{ref}{}
}
\value{
hypothesis matrix
}
\description{
Version of Ldiff used in RDC
}

