% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Lrm}
\alias{Lrm}
\title{Construct hypothesis matrix to test repeated measures factor effects}
\usage{
Lrm(fit, nam, vals = 1:nrow(L.mu))
}
\arguments{
\item{fit}{}

\item{nam}{}

\item{vals}{}
}
\value{
hypothesis matrix
}
\description{
Construct hypothesis matrix to test repeated measures factor effects
}

