% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lfx.R
\name{M}
\alias{*.M}
\alias{<.factor}
\alias{<=.factor}
\alias{>.factor}
\alias{>=.factor}
\alias{M}
\alias{M.M}
\alias{M.default}
\alias{M.factor}
\alias{M.formula}
\title{Method to generate hypothesis matrices from formulas}
\usage{
M(x, ...)

\method{M}{formula}(form, ..., keep.intercept = FALSE)

\method{M}{factor}(x, base = NULL)

\method{M}{M}(x, ...)

\method{M}{default}(x, ...)

\method{*}{M}(e1, e2)

\method{<}{factor}(x, y)

\method{>}{factor}(x, y)

\method{<=}{factor}(x, y)

\method{>=}{factor}(x, y)
}
\arguments{
\item{...}{other arguments}

\item{form}{formula}

\item{keep.intercept}{(default: FALSE) if TRUE generate a column for the intercept term}
}
\description{
Interprets formulas to generate hypothesis matrices
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: formula method

\item \code{factor}: factor method

\item \code{M}: method for M object

\item \code{default}: default method

\item \code{M}: multiplication of M objects

\item \code{factor}: '<' method for factors

\item \code{factor}: '>' method for factors

\item \code{factor}: '<=' method for factors

\item \code{factor}: '>=' method for factors
}}

