% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{pal}
\alias{pal}
\title{Generate a palette of colours -- possibly superseded}
\usage{
pal(col = c("blue", "pink"), border = "light gray", ...)
}
\arguments{
\item{col}{colors to show}

\item{border}{(default 'light gray')}

\item{\dots}{}
}
\description{
Generate a palette of colours -- possibly superseded
}

