% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.R
\name{part}
\alias{part}
\title{Intersection and symmetric differences of two sets}
\usage{
part(A, B)
}
\arguments{
\item{A}{}

\item{B}{}
}
\description{
Intersection and symmetric differences of two sets
}

