% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Pmat}
\alias{Pmat}
\alias{Pmat_diffmat}
\title{Comparison hypothesis matrix}
\usage{
Pmat(f, type = c("factor", "raw", "mean", "(mean)", "<mean>", "II", "cen",
  "cent", "center", "centre", "<centre>", "<center>", "(center)", "(centre)",
  "III", "pairwise", "<others.m>", "<others.c>", "diff", "diffmean", "diffc",
  "diffcen", "diffcentre", "diffcentre"), all = FALSE, sep = "")

Pmat_diffmat(w)
}
\arguments{
\item{f}{a factor or otherwise a vector interpreted as the levels of a
factor}

\item{type}{one "factor","raw","mean","(mean)","<mean>","II",
"cen","cent","center","centre","<centre>" , "<center>" ,
"(center)","(centre)","III","pairwise",  "<others.m>", "<others.c>",
"diff", "diffmean","diffc","diffcen","diffcentre","diffcentre"}

\item{all}{for some values of \code{type}, indicates whether to use all
contrasts. e.g. \code{type = "pairwise"} will produce pairwise comparisons
in both directions if \code{all == TRUE}}

\item{sep}{can add spaces in constructed factor levels}
}
\description{
Comparison hypothesis matrix
}
\section{Functions}{
\itemize{
\item \code{Pmat_diffmat}: a utility function to generate a matrix to apply to contrasts to yield differences from
means of other levels.
}}

