% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{print.cat}
\alias{print.cat}
\title{Print method for 'cat' objects}
\usage{
\method{print}{cat}(object, ...)
}
\arguments{
\item{x}{}

\item{\dots}{}
}
\value{
invisible(x)
}
\description{
Print method for 'cat' objects
}

