% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd}
\alias{rnd}
\title{Round and format}
\usage{
rnd(x, digits, ...)
}
\arguments{
\item{x}{an object to be rounded and formatted}

\item{digits}{number of digits for rounding}

\item{...}{passed on to \code{\link{format}} function}
}
\description{
Round to a given number of significant digits for variability
}
\details{
Instead of rounding to a given number of significant digits, or rounding
to a given significance, 'rnd' rounds so that the variability in a set of
number is exhibited to a given number of digits.
}
\examples{
rnd(c(0.00001111111,0.000022222,0.000015), scientific = F)
rnd(c(0.00001111111,0.000022222,0.000015)+100, scientific = F)
round(c(0.00001111111,0.000022222,0.000015))
round(c(0.00001111111,0.000022222,0.000015)+100)
rnd(123)
rnd(123.45678901)
}
\seealso{
\code{\link{pfmt}}, \code{\link{rpfmt}}
}

