% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spidanew.R
\docType{package}
\name{spidanew}
\alias{spidanew}
\alias{spidanew-package}
\title{spidanew: New functions for the spida package}
\description{
The spida package is a collection of functions and datasets intended primarily
for longitudinal and hierarchical data analysis.
}
\section{Functions}{


\code{\link{long}} is an attempt to simplify the functions 'long' and 'tolong'
in spida which, in turn, are attempts at simplifying calls to
\code{\link{stats::reshape}}
}

