% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tran.R
\name{tr}
\alias{tr}
\title{Translate elements of a character vector}
\usage{
tr(x, from, to)
}
\arguments{
\item{x}{}

\item{from}{}

\item{to}{}
}
\description{
Changes the order of arguments in \code{\link{tran}}.
}

