% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tran.R
\name{tran}
\alias{tran}
\title{Translate elements of a vector}
\usage{
tran(from, to, x, tofactor = is.factor(x))
}
\arguments{
\item{from}{}

\item{to}{}

\item{x}{}

\item{tofactor}{}
}
\description{
This should return something like 'to' whenever all of 'x'
is in 'from'
Otherwise, if some of x is not modified then it needs to
either to change those to NAs or to leave them as is.
For a strict function, we can simply index to which will
leave it unchanged
which conflicts with my versions of tr, ergo this is called tran
}

