% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{Vcov}
\alias{Vcor}
\alias{Vcov}
\title{Generic 'vcov' extended to objects with a \code{getFix} method}
\usage{
Vcov(fit)

Vcor(fit)
}
\arguments{
\item{fit}{}
}
\value{
estimated variance covariance matrix of fixed effects
}
\description{
Generic 'vcov' extended to objects with a \code{getFix} method

Correlation matrix of fixed effects
}
\section{Functions}{
\itemize{
\item \code{Vcor}: correlation matrix of fixed effects
}}
\author{
GM
}

