% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yscs.R
\docType{package}
\name{yscs}
\alias{yscs}
\alias{yscs-package}
\title{yscs: New functions for the spida package}
\description{
The yscs package is a collection of functions and datasets intended primarily
for statistical consulting, with a particular emphasis on
longitudinal and hierarchical data analysis.
}
\section{Wald tests and linear hypothesis matrices}{

\itemize{
\item \code{\link{wald}} quick wald tests building hypothesis matrices from regular expressions
\item \code{\link{walddf}}
\item \code{\link{as.data.frame.wald}}
\item \code{\link{coef.wald}}
\item \code{\link{print.wald}}
\item \code{\link{Lfx}}  creates hypothesis matrices for derivatives and differences.
\item \code{\link{Lall}}
\item \code{\link{Lc}}
\item \code{\link{Lcall}}
\item \code{\link{Ldiff}}
\item \code{\link{Lequal}}
\item \code{\link{Lform}}
\item \code{\link{Lmat}}
\item \code{\link{Lmu}}
\item \code{\link{Lrm}}
\item \code{\link{M}}
}
}

\section{Multilevel data frames}{

\itemize{
\item \code{\link{long}} is an attempt to simplify the functions 'long' and 'tolong'
  in spida which, in turn, are attempts at simplifying calls to
\item \code{\link{stats::reshape}}
}
}

\section{Datasets}{

\itemize{
\item \code{\link{hsfull}}
\item \code{\link{iq}}
}
}

\section{Miscellaneous utility functions}{

\itemize{
\item \code{\link{disp}}
\item \code{\link{getData}}
\item \code{\link{getFactorNames}}
\item \code{\link{getFix}}
\item \code{\link{getX}}
\item \code{\link{gpanel.fit}}
\item \code{\link{grapes-less-grapes}}
\item \code{\link{labs}}
\item \code{\link{pch}}
\item \code{\link{pfmt}}
\item \code{\link{Pmat}}
\item \code{\link{print.cat}}
\item \code{\link{rnd}}
\item \code{\link{rpfmt}}
\item \code{\link{run}}
\item \code{\link{Vcov}}
\item \code{\link{xlevels}}
}
}

